backend/pyproject.toml
[project]
name = "mcp-scanner"
version = "0.1.0"
description = "Enterprise MCP Security Scanner"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "gitpython>=3.1.0",
    "weasyprint>=63.0",
    "python-multipart>=0.0.18",
    "yara-python>=4.5.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.24.0",
    "tree-sitter-javascript>=0.24.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-go>=0.24.0",
    "cyclonedx-python-lib>=11.0.0",
    "packageurl-python>=0.15.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mcp_scanner = ["data/*.json", "data/yara_rules/*.yar", "data/prompts/*.txt"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
